The search API in Drupal 7.4 changed significantly, and these changes make our search interfaces compatible with those changes. Proof of test on localhost: https://skitch.com/richard-alexander-allen/fjhnx/photos-and-video-time-for-kids
In apachesolr.module beta-6 version the Facets are no longer declared in that module and are moved to the Facet API. This module comes with D7.4 and needs to be enabled. Contexts need to be modified, and Features rebuilt. Features can also export the facets in D7.4 using this interface: https://skitch.com/richard-alexander-allen/fjhba/features-time-for-kids