re #621: A link for Internal Use Only is appearing on an ATW Destination page
This is a taxonomy term field that has the name and label "Internal Use Only." It was included in the full-node display of the ATW Destination content type, instead of being hidden. The field and its label are now hidden and feature tfk_atw_feature has been recreated/updated to include the field change.