re #606: Homework Helper A+ papers
The first issue (Script font in FF) could not be reproduced either on Mac or Windows.

For the second issue, the Ajax form is being prefilled with different user's data if there are several users active at the same time on the same form. Change was made to clear out the current user's form data their first time through, to avoid this.