re #628: Minisite flagged to the Homepage secondary touts is not displayed
Minisite flagging for secondary news touts (homepage, news, and kid reporter) is not working - we will address it in a later release.
In the meantime, this update removes the availability of the Secondary Touts under News, Secondary Touts in Kid Reporter, and Secondary Touts on Homepage flags from Minisites content type.