re #611, #617: Subscribe & Renew links
Custom code in several feature_misc .inc files was causing problems with the feature, including the addition of an extra menu item. It also did not properly refresh the menu links for Subscribe & Renew when the LUCIE Settings were changed.

Removed the hand-coded bits from the feature's .inc files, and added a submit callback on the LUCIE Settings form to modify all menu items and clear the cache to force the menus, as well as the theme registry (which updates the S&R link on the login form), to be rebuilt.